Moved TranslogCheckedContainer to org.opensearch.index.translog
