--- server/src/main/java/org/opensearch/index/translog/checked/package-info.java
@@ -1,10 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- */
-
-/** Contains checksum related utilities for translog files */
-package org.opensearch.index.translog.checked;
